backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachelib==0.12.0
chardet==5.2.0
click==8.1.7
fastapi==0.109.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Session==0.6.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psycopg2==2.9.9
pydantic==2.6.0
pydantic_core==2.16.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.35.1
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1